Expand mod skeleton to accomodate larger mod
The mod skeleton from the previous commit was adapted from my previous mod Wrench in the Works, which fits into two files (3 for 1.10). Magicka, on the other hand, is going to be a much more expansive mod, so it's skeleton needed a bit of expansion to accomodate.

I've also added the US localization file and main creative tab for later use.